chore: move ui folder to src folder
