+++ src/app/page.js
@@ -1,5 +1,5 @@
-import Countdown from "../ui/countdown";
-import Accordion from "../ui/faq.jsx";
+import Countdown from "./ui/countdown";
+import Accordion from "./ui/faq.jsx";
 export default function Home() {
   return (
     <div>
